vspfilter: Add a new plugin converting colorspace and video size with VSP1 V4L2
This patch adds a new plugin that can use the color conversion and
scaling with the hardware acceleration on R-Car SoC. This function is
available through the V4L2 driver I/F and the userptr method is used for
the buffer passing between the GStreamer plugin and the V4L2 in order to
do zero-copy processing.
The dmabuf method is also available just for the output port, however
it performs on the linux kernel 3.7 rc7 and later.

Change-Id: I969e0f9b9c589d0be50de6f613f4514db9fa7289
Signed-off-by: Kazunori Kobayashi <kkobayas@igel.co.jp>