playsink: Move the queue plugin insertion after the videoconvert plugin in the pipeline
When the zero-copy buffer passing between decoder and videoconvert plugins
performs, the decoding and the conversion processing may not synchronize
if a queue plugin is inserted between them. To avoid this, the queue plugin
should be inserted after the videoconvert plugin because the synchronization
is ensured by the buffer management of the bufferpool in videosink plugin.

When the pool's buffers are used, this malfunction does not occur
because it is ensured that the each buffer once returns to the pool
after the use of the buffer and the next buffer acquisition will
perform after that.

Change-Id: Id944de58946509e3ba09952301fda6e8b21f19a0
Signed-off-by: Kazunori Kobayashi <kkobayas@igel.co.jp>