Prevent the NEON check in configure from passing under aarch64.
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package
"gst-plugins-base" cannot be built for this architecture because of build
errors.

This bug was originally obtained for open-source version of package
"gst-plugins-base", bug #712367.

Here is the commit message for patch that was written to fix this bug:

The test verifies that the NEON C intrinsics work, but the rest of the
codebase uses lots of direct ARMv7 NEON assembly. The same intrinsics
work in A64, but the assembly is slightly different.

Prevent the check from passing so that we don't use this where it won't
work.

Patch originally by William Grant <wgrant@ubuntu.com>

For more details, see
https://bugzilla.gnome.org/show_bug.cgi?id=712367

Change-Id: Ieff5d7e5d1b4d2842d857e5dbdf04ba67736eb0d
Signed-off-by: Ilya Palachev <i.palachev@samsung.com>